move finding an attribute apart of its parser
