+++ src/Domain/Objects/AttributeParser/index.js
@@ -1,20 +1,4 @@
-const DEPENDENCIES = {
-  attributes: [
-    require('./Attributes/StringAttribute'),
-    require('./Attributes/BooleanAttribute'),
-    require('./Attributes/IntegerAttribute'),
-    require('./Attributes/DateAttribute'),
-    require('./Attributes/FloatAttribute'),
-    require('./Attributes/ObjectAttribute'),
-    require('./Attributes/JSONAttribute'),
-    require('./Attributes/ArrayAttribute')
-  ],
-  defaultAttribute: require('./Attributes/TokenAttribute')
-}
-
-function isMatch (parser) {
-  return parser.isMatch(this.type)
-}
+const findAttribute = require('./findAttribute')
 
 function AttributeParser (data, injection) {
   if (typeof data === 'string') {
@@ -23,12 +7,7 @@ function AttributeParser (data, injection) {
     return AttributeParser(typedAttribute, injection)
   }
 
-  const { attributes, defaultAttribute } = Object.assign({}, DEPENDENCIES, injection)
-  const { type } = data
-
-  const attribute = attributes.find(isMatch, { type })
-
-  const Attribute = attribute || defaultAttribute
+  const Attribute = findAttribute({ type: data.type }, injection)
 
   return new Attribute(data)
 }
